refactor: move extension to be under ext package as it is specified for outside communication
