+++ ext/extension/manifest.go
