+++ ext/extension/manifest_test.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/stretchr/testify/assert"
 
-	"github.com/odpf/optimus/extension"
+	"github.com/odpf/optimus/ext/extension"
 )
 
 func TestLoadManifest(t *testing.T) {
